fix build warning

+++ engine/src/lua_cpplib/lnetwork_mgr.hpp
@@ -312,8 +312,11 @@ public:
     bool connect_new(uint32_t conn_id, int32_t ecode);
     bool accept_new(uint32_t conn_id, class Socket *new_sk);
 
-    /* 自动转发 */
-    bool cs_dispatch(int32_t cmd, const class Socket *src_sk, const char *ctx,
+    /**
+     * 把客户端数据包转发给另一服务器
+     * 这个函数如果返回false，则会将协议在当前进程派发
+     */
+    bool cs_dispatch(uint16_t cmd, const class Socket *src_sk, const char *ctx,
                      size_t size) const;
 
 private:
